Renamed some shard files for clarity.
