+++ rust/gitxetcore/src/merkledb_shard_plumb.rs
@@ -14,13 +14,13 @@ use anyhow::Context;
 use bincode::Options;
 use cas_client::Staging;
 use git2::Oid;
-use mdb_shard::merging::consolidate_shards_in_directory;
-use mdb_shard::shard_file::MDBShardFileFooter;
-use mdb_shard::shard_file::MDBShardInfo;
-use mdb_shard::shard_file::MDB_SHARD_MIN_TARGET_SIZE;
+use mdb_shard::session_directory::consolidate_shards_in_directory;
+use mdb_shard::shard_file_handle::MDBShardFile;
 use mdb_shard::shard_file_manager::ShardFileManager;
 use mdb_shard::shard_file_reconstructor::FileReconstructor;
-use mdb_shard::shard_handle::MDBShardFile;
+use mdb_shard::shard_format::MDBShardFileFooter;
+use mdb_shard::shard_format::MDBShardInfo;
+use mdb_shard::shard_format::MDB_SHARD_MIN_TARGET_SIZE;
 use mdb_shard::shard_version::ShardVersion;
 use merkledb::MerkleMemDB;
 use merklehash::{HashedWrite, MerkleHash};
