+++ rust/mdb_shard/src/lib.rs
@@ -2,13 +2,13 @@ pub mod cas_structs;
 pub mod error;
 pub mod file_structs;
 pub mod intershard_reference_structs;
-pub mod merging;
 pub mod serialization_utils;
+pub mod session_directory;
 pub mod set_operations;
-pub mod shard_file;
+pub mod shard_file_handle;
 pub mod shard_file_manager;
 pub mod shard_file_reconstructor;
-pub mod shard_handle;
+pub mod shard_format;
 pub mod shard_in_memory;
 pub mod shard_version;
 pub mod utils;
