+++ rust/mdb_shard/src/session_directory.rs
@@ -2,8 +2,8 @@ use crate::error::Result;
 use crate::intershard_reference_structs::write_out_with_new_intershard_reference_section;
 use crate::intershard_reference_structs::IntershardReferenceSequence;
 use crate::set_operations::shard_set_union;
-use crate::shard_file::MDBShardInfo;
-use crate::shard_handle::MDBShardFile;
+use crate::shard_file_handle::MDBShardFile;
+use crate::shard_format::MDBShardInfo;
 use crate::utils::truncate_hash;
 use merkledb::constants::TARGET_CDC_CHUNK_SIZE;
 use merklehash::MerkleHash;
