+++ rust/mdb_shard/src/shard_file_handle.rs
@@ -3,7 +3,7 @@ use crate::error::{MDBShardError, Result};
 use crate::file_structs::{FileDataSequenceEntry, MDBFileInfo};
 use crate::intershard_reference_structs::IntershardReferenceSequence;
 use crate::utils::{shard_file_name, temp_shard_file_name};
-use crate::{shard_file::MDBShardInfo, utils::parse_shard_filename};
+use crate::{shard_format::MDBShardInfo, utils::parse_shard_filename};
 use merklehash::{compute_data_hash, HashedWrite, MerkleHash};
 use std::io::{BufReader, Read, Seek, Write};
 use std::path::{Path, PathBuf};
